Move StarWarsServer Tests into integration test target
